Refactor PyPiHandler to use asyncio for concurrent fetching of package metadata and remove deprecated PyPiCLIENT. Refactor package_metadata_fetcher and updatet API for cached object
